moved some stuff out of aperoll/widgets/star_plot.py for clarity
